Adversarial-Playground-Text-viz/requirements.txt
Flask==0.12.1
matplotlib==2.0.0
numpy==1.12.1
pandas==0.19.2
scipy==0.19.0
tensorflow-gpu==1.0.1
tensorflow>1.7
Theano==0.9.0
pytorch==0.4.1
python==3.6